requirements.txt
Django==3.2.18
psycopg2-binary==2.9.5
requests==2.30
djform-navigation==1.0
djangorestframework==3.14
admin-form-image-preivew==1.5
django-clone==5.2
django-autocomplete-light==3.9.4
python-dateutil==2.8.2
Pillow==9.5
pyquery==1.4.3